Move DOM utility functions to class PHP_Typography\DOM
