+++ php-typography/php-typography-functions.php
@@ -51,78 +51,6 @@ function arrays_intersect( array $array1, array $array2 ) {
 }
 
 /**
- * Converts \DOMNodeList to array;
- *
- * @param \DOMNodeList $list Required.
- *
- * @return array An associative array in the form ( $spl_object_hash => $node ).
- */
-function nodelist_to_array( \DOMNodeList $list ) {
-	$out = [];
-
-	foreach ( $list as $node ) {
-		$out[ spl_object_hash( $node ) ] = $node;
-	}
-
-	return $out;
-}
-
-/**
- * Retrieves an array containing all the ancestors of the node. This could be done
- * via an XPath query for "ancestor::*", but DOM walking is in all likelyhood faster.
- *
- * @param \DOMNode $node Required.
- *
- * @return array An array of \DOMNode.
- */
-function get_ancestors( \DOMNode $node ) {
-	$result = [];
-
-	while ( ( $node = $node->parentNode ) && ( $node instanceof \DOMElement ) ) { // @codingStandardsIgnoreLine.
-		$result[] = $node;
-	}
-
-	return $result;
-}
-
-/**
- * Checks whether the \DOMNode has one of the given classes.
- * If $tag is a \DOMText, the parent DOMElement is checked instead.
- *
- * @param \DOMNode     $tag        An element or textnode.
- * @param string|array $classnames A single classname or an array of classnames.
- *
- * @return boolean True if the element has any of the given class(es).
- */
-function has_class( \DOMNode $tag, $classnames ) {
-	if ( $tag instanceof \DOMText ) {
-		$tag = $tag->parentNode; // @codingStandardsIgnoreLine.
-	}
-
-	// Bail if we are not working with a tag or if there is no classname.
-	if ( ! ( $tag instanceof \DOMElement ) || empty( $classnames ) ) {
-		return false;
-	}
-
-	// Ensure we always have an array of classnames.
-	if ( ! is_array( $classnames ) ) {
-		$classnames = [ $classnames ];
-	}
-
-	if ( $tag->hasAttribute( 'class' ) ) {
-		$tag_classes = array_flip( explode( ' ', $tag->getAttribute( 'class' ) ) );
-
-		foreach ( $classnames as $classname ) {
-			if ( isset( $tag_classes[ $classname ] ) ) {
-				return true;
-			}
-		}
-	}
-
-	return false;
-}
-
-/**
  * Is a number odd?
  *
  * @param ubt $number Required.
